refactor: use splitArray from web-utility and commander-jsx to add optional command
