+++ package.json
@@ -4,8 +4,10 @@
   "type": "module",
   "source": "src/index.ts",
   "dependencies": {
+    "commander-jsx": "^0.6.9",
     "jsdom": "^22.1.0",
     "koapache": "^2.2.2",
+    "web-utility": "^4.1.3",
     "zx": "^7.2.3"
   },
   "devDependencies": {
