+++ tsconfig.json
@@ -3,7 +3,9 @@
     "target": "ES2017",
     "module": "NodeNext",
     "moduleResolution": "NodeNext",
+    "jsx": "react",
+    "jsxFactory": "createCommand",
     "outDir": "dist/"
   },
-  "include": ["src/**/*.ts"]
+  "include": ["src/index.tsx", "src/**/*.tsx"]
 }
